We won't use "plural" anymore, refactor
