+++ src/pages/student/index.tsx
