+++ src/pages/student/training/list/page.tsx
