+++ src/routes.tsx
@@ -5,7 +5,7 @@ import { Router, Route, IndexRoute } from 'react-router';
 import { App } from './app';
 import { GeneralRoutes, NotFoundPage } from './pages/general'
 import { AdminRoutes } from './pages/admin';
-import { StudentsRoutes } from './pages/students';
+import { StudentsRoutes } from './pages/student';
 
 export const AppRoutes = (
     <Route  path="/" component={App} >
